Add match_type to allow type to be supplied. (#9)
Add some internal rules to allow a type to be supplied at the end of the comp. This allows
most of the collection types to be used. Additionally, update some small things (add if
cond) for two cases)

In addition, add test files for each case and populate with basic tests.